debugging deployment
