+++ package.json
@@ -1,6 +1,6 @@
 {
   "description": "Feel free to delete this and start from scratch.",
-  "main": "index.js",
+  "main": "app.js",
   "type": "module",
   "scripts": {
     "dev": "nodemon -r dotenv/config app.js",
